+++ glibc/string/mempcpy.c
@@ -1,5 +1,5 @@
 /**
- * char *stpcpy(char *dst, const char *src);
+ * void *mempcpy(void dest[n], const void src[n], size_t n);
  */
 #include <string.h>
 #include <stdio.h>
@@ -9,9 +9,7 @@ int main(void)
 	char buffer[10] = { "123456789" };
 	char *to = buffer;
 
-	to = stpcpy(to, "foo");
-	to = stpcpy(to, "bar");
-	to = stpcpy(to, "1");
+	to = mempcpy(to, "foo", 3);
 
 	puts(to);
 	puts(buffer);